PKI acceptance tests stabilized
