+++ ui/tests/acceptance/pki/pki-cross-sign-test.js
@@ -9,10 +9,10 @@ import { setupApplicationTest } from 'vault/tests/helpers';
 import { v4 as uuidv4 } from 'uuid';
 
 import authPage from 'vault/tests/pages/auth';
-import enablePage from 'vault/tests/pages/settings/mount-secret-backend';
-import { runCommands } from 'vault/tests/helpers/pki/pki-run-commands';
 import { SELECTORS } from 'vault/tests/helpers/pki/pki-issuer-cross-sign';
 import { verifyCertificates } from 'vault/utils/parse-pki-cert';
+import { deleteEngineCmd, mountEngineCmd, runCmd } from 'vault/tests/helpers/commands';
+
 module('Acceptance | pki/pki cross sign', function (hooks) {
   setupApplicationTest(hooks);
 
@@ -24,10 +24,11 @@ module('Acceptance | pki/pki cross sign', function (hooks) {
     this.intMountPath = `intermediate-mount-${uuidv4()}`; // first input box in cross-signing page
     this.intIssuerName = 'my-intermediate-issuer'; // second input box in cross-signing page
     this.newlySignedIssuer = 'my-newly-signed-int'; // third input
-    await enablePage.enable('pki', this.parentMountPath);
-    await enablePage.enable('pki', this.intMountPath);
-
-    await runCommands([
+    await runCmd(
+      [mountEngineCmd('pki', this.parentMountPath), mountEngineCmd('pki', this.intMountPath)],
+      false
+    );
+    await runCmd([
       `write "${this.parentMountPath}/root/generate/internal" common_name="Long-Lived Root X1" ttl=8960h issuer_name="${this.oldParentIssuerName}"`,
       `write "${this.parentMountPath}/root/generate/internal" common_name="Long-Lived Root X2" ttl=8960h issuer_name="${this.parentIssuerName}"`,
       `write "${this.parentMountPath}/config/issuers" default="${this.parentIssuerName}"`,
@@ -36,8 +37,7 @@ module('Acceptance | pki/pki cross sign', function (hooks) {
 
   hooks.afterEach(async function () {
     // Cleanup engine
-    await runCommands([`delete sys/mounts/${this.intMountPath}`]);
-    await runCommands([`delete sys/mounts/${this.parentMountPath}`]);
+    await runCmd([deleteEngineCmd(this.intMountPath), deleteEngineCmd(this.parentMountPath)]);
   });
 
   test('it cross-signs an issuer', async function (assert) {
@@ -50,6 +50,7 @@ module('Acceptance | pki/pki cross sign', function (hooks) {
     const csr = find(SELECTORS.copyButton('CSR')).getAttribute('data-test-copy-button');
     await visit(`vault/secrets/${this.parentMountPath}/pki/issuers/${this.oldParentIssuerName}/sign`);
     assert.strictEqual(currentRouteName(), 'vault.cluster.secrets.backend.pki.issuers.issuer.sign');
+
     await fillIn(SELECTORS.inputByName('csr'), csr);
     await fillIn(SELECTORS.inputByName('format'), 'pem_bundle');
     await click('[data-test-pki-sign-intermediate-save]');
@@ -89,7 +90,7 @@ module('Acceptance | pki/pki cross sign', function (hooks) {
 
     // verify cross-sign was accurate by creating a role to issue a leaf certificate
     const myRole = 'some-role';
-    await runCommands([
+    await runCmd([
       `write ${this.intMountPath}/roles/${myRole} \
     issuer_ref=${this.newlySignedIssuer}\
     allow_any_name=true \
