+++ ui/tests/helpers/pki.js
@@ -21,20 +21,3 @@ export const STANDARD_META = {
   currentPage: 1,
   pageSize: 100,
 };
-
-export const clearPkiRecords = (store) => {
-  // Clears pki-related data and capabilities so that admin
-  // capabilities from setup don't rollover in permissions tests
-  store.unloadAll('pki/issuer');
-  store.unloadAll('pki/action');
-  store.unloadAll('pki/config/acme');
-  store.unloadAll('pki/certificate/generate');
-  store.unloadAll('pki/certificate/sign');
-  store.unloadAll('pki/config/cluster');
-  store.unloadAll('pki/key');
-  store.unloadAll('pki/role');
-  store.unloadAll('pki/sign-intermediate');
-  store.unloadAll('pki/tidy');
-  store.unloadAll('pki/config/urls');
-  store.unloadAll('capabilities');
-};
